LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/worker-src-none/worker-import-data.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=07ddee9e-3536-431a-8039-f50a4a7bbd7b&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=02b23001-8e30-4660-80c1-05fa1cbf52a9&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=f9bf5a8b-a2be-4811-a18a-becb159d2742&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?action=purge&key=0b029e67-35bb-4c7f-8017-4206f027cb68&path=%2Fmixed-content

Harness Error (TIMEOUT), message = null

PASS Content Security Policy: Expects blocked for worker-import-data to cross-http origin and keep-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import-data to cross-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import-data to cross-http origin and no-redirect redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import-data to cross-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import-data to cross-http origin and swap-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import-data to cross-http origin and swap-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for worker-import-data to same-http origin and keep-origin redirection from http context. assert_unreached: main promise resolved unexpectedly Reached unreachable code
FAIL Content Security Policy: Expects blocked for worker-import-data to same-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for worker-import-data to same-http origin and no-redirect redirection from http context. assert_unreached: main promise resolved unexpectedly Reached unreachable code
FAIL Content Security Policy: Expects blocked for worker-import-data to same-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
TIMEOUT Content Security Policy: Expects blocked for worker-import-data to same-http origin and swap-origin redirection from http context. Test timed out
NOTRUN Content Security Policy: Expects blocked for worker-import-data to same-http origin and swap-origin redirection from http context.: securitypolicyviolation
